+++ apps/web/src/app/features/home/components/index.ts
@@ -1,2 +1,6 @@
+export * from './ChannelList';
+export * from './CreateChannelDialog';
+export * from './CreatePlaceDialog';
+export * from './PlaceList';
 export * from './ProfileEditModal';
 export * from './CloudSessionSheet';